% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ang2rad}
\alias{ang2rad}
\title{Utilities for manipulation}
\usage{
ang2rad(angle)
}
\arguments{
\item{angle}{Angle \code{angle}}
}
\value{
radian
}
\description{
Convert angle to radian
}
\examples{
angles <- 0:180
radian <- ang2rad(angle = angles)
}
