% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut.R
\name{donut}
\alias{donut}
\title{data-driven donuts}
\usage{
donut(.data, .var, delay = 0)
}
\description{
data-driven donuts
}
\examples{
set.seed(1004)
.data <- data.frame(
				G1 = sample(LETTERS[1:5],100,TRUE),
			G2 = sample(LETTERS[1:3],100,TRUE)
		)
.var <- colnames(.data)[1]
donut(.data, .var)
}
