% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut-prop2arc.R
\name{prop2ang}
\alias{prop2ang}
\title{Proportion to Angle}
\usage{
prop2ang(prop)
}
\arguments{
\item{prop}{Proportion \code{prop}}
}
\description{
Convert proportion to angles for arc
}
\examples{
prop <- c(A = 0.1,B = 0.3, C = 0.6)
prop2ang(prop)
}
