% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigo.R
\name{Sin Cos Tan}
\alias{Sin Cos Tan}
\alias{Sin}
\alias{Cos}
\alias{Tan}
\title{Trigonometric functions}
\usage{
Sin(angle, digit = 10)

Cos(angle, digit = 10)

Tan(angle, digit = 10)
}
\arguments{
\item{angle}{Angle \code{angle}}

\item{digit}{Digit, default is 10 \code{digit}}
}
\description{
Use angle as input instead of radian, Sin, Cos, Tan
}
\examples{
Sin(90)
Cos(90)
Tan(45)
}
