% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donut-arc.R
\name{draw_arc}
\alias{draw_arc}
\title{Draw arc without padding}
\usage{
draw_arc(x, y, s.ang, e.ang, stroke = "skyblue", width = 5, height = 5,
  stroke.size = 4, stroke.position = "out")
}
\description{
Need to add padding
}
